+++ src/fr/maxlego08/spawner/buttons/virtual/InfoButton.java
@@ -1,0 +1,33 @@
+package fr.maxlego08.spawner.buttons.virtual;
+
+import fr.maxlego08.menu.api.utils.Placeholders;
+import fr.maxlego08.menu.button.ZButton;
+import fr.maxlego08.spawner.SpawnerManager;
+import fr.maxlego08.spawner.SpawnerPlugin;
+import fr.maxlego08.spawner.api.Spawner;
+import fr.maxlego08.spawner.api.utils.PlayerSpawner;
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.plugin.Plugin;
+
+public class InfoButton extends ZButton {
+
+    private final SpawnerPlugin plugin;
+
+    public InfoButton(Plugin plugin) {
+        this.plugin = (SpawnerPlugin) plugin;
+    }
+
+    @Override
+    public ItemStack getCustomItemStack(Player player) {
+
+        Placeholders placeholders = new Placeholders();
+
+        SpawnerManager manager = this.plugin.getManager();
+        PlayerSpawner playerSpawner = manager.getPlayerSpawners().get(player.getUniqueId());
+        Spawner spawner = playerSpawner == null ? null : playerSpawner.getVirtualSpawner() == null ? null : playerSpawner.getVirtualSpawner();
+        manager.registerPlaceholders(placeholders, spawner);
+
+        return getItemStack().build(player, false, placeholders);
+    }
+}