+++ src/fr/maxlego08/spawner/dto/ItemDTO.java
@@ -1,0 +1,6 @@
+package fr.maxlego08.spawner.dto;
+
+import java.util.UUID;
+
+public record ItemDTO(UUID spawnerId, String itemStack, long amount) {
+}